Work in progress refactor FlatSpec test into its own module.:wq
