+++ build.sbt
@@ -17,18 +17,22 @@ lazy val scalatest = ScalatestBuild.scalatest
 lazy val scalatestTest = ScalatestBuild.scalatestTest
 lazy val scalatestDiagramsTest = ScalatestBuild.scalatestDiagramsTest
 lazy val scalatestFeatureSpecTest = ScalatestBuild.scalatestFeatureSpecTest
+lazy val scalatestFlatSpecTest = ScalatestBuild.scalatestFlatSpecTest
 lazy val scalatestJS = ScalatestBuild.scalatestJS
 lazy val scalatestNative = ScalatestBuild.scalatestNative
 lazy val scalatestDotty = ScalatestBuild.scalatestDotty
 lazy val scalatestTestJS = ScalatestBuild.scalatestTestJS
 lazy val scalatestDiagramsTestJS = ScalatestBuild.scalatestDiagramsTestJS
 lazy val scalatestFeatureSpecTestJS = ScalatestBuild.scalatestFeatureSpecTestJS
+lazy val scalatestFlatSpecTestJS = ScalatestBuild.scalatestFlatSpecTestJS
 lazy val scalatestTestNative = ScalatestBuild.scalatestTestNative
 lazy val scalatestDiagramsTestNative = ScalatestBuild.scalatestDiagramsTestNative
 lazy val scalatestFeatureSpecTestNative = ScalatestBuild.scalatestFeatureSpecTestNative
+lazy val scalatestFlatSpecTestNative = ScalatestBuild.scalatestFlatSpecTestNative
 lazy val scalatestTestDotty = ScalatestBuild.scalatestTestDotty
 lazy val scalatestDiagramsTestDotty = ScalatestBuild.scalatestDiagramsTestDotty
 lazy val scalatestFeatureSpecTestDotty = ScalatestBuild.scalatestFeatureSpecTestDotty
+lazy val scalatestFlatSpecTestDotty = ScalatestBuild.scalatestFlatSpecTestDotty
 lazy val scalatestApp = ScalatestBuild.scalatestApp
 lazy val scalatestAppJS = ScalatestBuild.scalatestAppJS
 lazy val scalatestAppNative = ScalatestBuild.scalatestAppNative
