+++ jvm/flatspec-test/src/test/scala/org/scalatest/flatspec/FixtureAsyncFlatSpecSpec2.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest.fixture
+package org.scalatest.flatspec
 
 import scala.concurrent.{ExecutionContext, Promise, Future}
 import org.scalatest._
@@ -25,13 +25,13 @@ import scala.util.Success
 import org.scalatest
 import org.scalatest.flatspec
 
-class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
+class FixtureAsyncFlatSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
 
-  describe("AsyncFlatSpecLike") {
+  describe("AsyncFlatSpec") {
 
     it("can be used for tests that return Future under parallel async test execution") {
 
-      class ExampleSpec extends flatspec.FixtureAsyncFlatSpecLike with ParallelTestExecution {
+      class ExampleSpec extends flatspec.FixtureAsyncFlatSpec with ParallelTestExecution {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -94,7 +94,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("can be used for tests that did not return Future under parallel async test execution") {
 
-      class ExampleSpec extends flatspec.FixtureAsyncFlatSpecLike with ParallelTestExecution {
+      class ExampleSpec extends flatspec.FixtureAsyncFlatSpec with ParallelTestExecution {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -149,7 +149,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
       @volatile var count = 0
 
-      class ExampleSpec extends flatspec.FixtureAsyncFlatSpecLike {
+      class ExampleSpec extends flatspec.FixtureAsyncFlatSpec {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -196,8 +196,8 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
       @volatile var count = 0
 
-      class ExampleSpec extends flatspec.FixtureAsyncFlatSpecLike {
-        
+      class ExampleSpec extends flatspec.FixtureAsyncFlatSpec {
+
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -241,7 +241,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
       var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends flatspec.FixtureAsyncFlatSpecLike {
+      class ExampleSpec extends flatspec.FixtureAsyncFlatSpec {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -288,7 +288,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
       @volatile var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends flatspec.FixtureAsyncFlatSpecLike {
+      class ExampleSpec extends flatspec.FixtureAsyncFlatSpec {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -351,7 +351,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("should not run out of stack space with nested futures when using SerialExecutionContext") {
 
-      class ExampleSpec extends flatspec.FixtureAsyncFlatSpecLike {
+      class ExampleSpec extends flatspec.FixtureAsyncFlatSpec {
 
         // Note we get a StackOverflowError with the following execution
         // context.
@@ -387,7 +387,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("should run tests that returns Future and report their result in serial") {
 
-      class ExampleSpec extends flatspec.FixtureAsyncFlatSpecLike {
+      class ExampleSpec extends flatspec.FixtureAsyncFlatSpec {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -438,7 +438,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("should run tests that does not return Future and report their result in serial") {
 
-      class ExampleSpec extends flatspec.FixtureAsyncFlatSpecLike {
+      class ExampleSpec extends flatspec.FixtureAsyncFlatSpec {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -482,7 +482,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in main spec body") {
-      class MySuite extends flatspec.FixtureAsyncFlatSpecLike  {
+      class MySuite extends flatspec.FixtureAsyncFlatSpec  {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -507,7 +507,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in test body") {
-      class MySuite extends flatspec.FixtureAsyncFlatSpecLike  {
+      class MySuite extends flatspec.FixtureAsyncFlatSpec  {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -539,7 +539,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in Future returned by scenario body") {
-      class MySuite extends flatspec.FixtureAsyncFlatSpecLike  {
+      class MySuite extends flatspec.FixtureAsyncFlatSpec  {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -573,7 +573,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in main spec body") {
-      class MySuite extends flatspec.FixtureAsyncFlatSpecLike  {
+      class MySuite extends flatspec.FixtureAsyncFlatSpec  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -596,7 +596,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in test body") {
-      class MySuite extends flatspec.FixtureAsyncFlatSpecLike  {
+      class MySuite extends flatspec.FixtureAsyncFlatSpec  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -619,7 +619,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in Future returned by test body") {
-      class MySuite extends flatspec.FixtureAsyncFlatSpecLike  {
+      class MySuite extends flatspec.FixtureAsyncFlatSpec  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -644,7 +644,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in main spec body") {
-      class MySuite extends flatspec.FixtureAsyncFlatSpecLike  {
+      class MySuite extends flatspec.FixtureAsyncFlatSpec  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -667,7 +667,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in test body") {
-      class MySuite extends flatspec.FixtureAsyncFlatSpecLike  {
+      class MySuite extends flatspec.FixtureAsyncFlatSpec  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -690,7 +690,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in Future returned by test body") {
-      class MySuite extends flatspec.FixtureAsyncFlatSpecLike  {
+      class MySuite extends flatspec.FixtureAsyncFlatSpec  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -715,7 +715,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in main spec body") {
-      class MySuite extends flatspec.FixtureAsyncFlatSpecLike  {
+      class MySuite extends flatspec.FixtureAsyncFlatSpec  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -738,7 +738,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in test body") {
-      class MySuite extends flatspec.FixtureAsyncFlatSpecLike  {
+      class MySuite extends flatspec.FixtureAsyncFlatSpec  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -768,7 +768,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in Future returned by scenario body") {
-      class MySuite extends flatspec.FixtureAsyncFlatSpecLike  {
+      class MySuite extends flatspec.FixtureAsyncFlatSpec  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -800,7 +800,7 @@ class AsyncFlatSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should allow other execution context to be used") {
-      class TestSpec extends flatspec.FixtureAsyncFlatSpecLike {
+      class TestSpec extends flatspec.FixtureAsyncFlatSpec {
         // SKIP-SCALATESTJS,NATIVE-START
         override implicit val executionContext = scala.concurrent.ExecutionContext.Implicits.global
         // SKIP-SCALATESTJS,NATIVE-END
