+++ project/scalatest.scala
@@ -235,6 +235,7 @@ object ScalatestBuild extends BuildCommons with DottyBuild with NativeBuild with
       "-m", "org.scalatest.expectations",
       "-m", "org.scalatest.diagrams",
       "-m", "org.scalatest.featurespec",
+      "-m", "org.scalatest.flatspec",
       "-oDIF",
       "-W", "120", "60",
       "-h", "target/html",
@@ -378,7 +379,14 @@ object ScalatestBuild extends BuildCommons with DottyBuild with NativeBuild with
     .settings(sharedTestSettings: _*)
     .settings(
       projectTitle := "ScalaTest FeatureSpec Test"
-    ).dependsOn(commonTest % "test")  
+    ).dependsOn(commonTest % "test")
+
+  lazy val scalatestFlatSpecTest = Project("scalatestFlatSpecTest", file("jvm/flatspec-test"))
+    .settings(sharedSettings: _*)
+    .settings(sharedTestSettings: _*)
+    .settings(
+      projectTitle := "ScalaTest FlatSpec Test"
+    ).dependsOn(commonTest % "test")    
 
   lazy val scalatestApp = Project("scalatestApp", file("scalatest-app"))
     .enablePlugins(SbtOsgi)
